+++ src/lib/client-layout.tsx
@@ -1,12 +1,14 @@
 "use client";
 
-import { FavoritesProvider } from "@/app/Providers/FavoritesContext";
+import { FavoritesProvider } from "@/app/providers/FavoritesContext";
 import GlobalStyle from "@/styles/globalStyle";
 import theme from "@/styles/theme";
 import { ReactNode } from "react";
 import { ThemeProvider } from "styled-components";
 
-export default function ClientLayout({ children }: Readonly<{ children: ReactNode }>) {
+export default function ClientLayout({
+  children,
+}: Readonly<{ children: ReactNode }>) {
   return (
     <ThemeProvider theme={theme}>
       <FavoritesProvider>
