+++ src/lib/styled-components-registry.tsx
@@ -1,4 +1,4 @@
-'use client'
+"use client";
 
 import React, { useState } from "react";
 import { useServerInsertedHTML } from "next/navigation";
