+++ tsconfig.json
@@ -20,8 +20,12 @@
     ],
     "paths": {
       "@/*": ["./src/*"],
+      "@types/*": ["./src/@types/*"],
+      "@api/*": ["./src/app/api/*"],
       "@components/*" :  [ "./src/app/_components/*" ] , 
       "@styled/*" :  [ "./src/app/_components/styled/*" ] , 
+      "@hooks/*" :  [ "./src/app/hooks/*" ] , 
+      "@providers/*" :  [ "./src/app/providers/*" ] , 
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
